12-pagina-di-conferma (#17)
* Crea la cartella private/data, se non esiste

* Fix #2

* hotfix aggiornata la versione di js per il controllo di linting

* Fix issue #14-Template-engine

* 12-pagina-di-conferma

* fix #12-pagina-di-conferma
