+++ src/email-sent.js
@@ -19,7 +19,6 @@ module.exports = function (request, h) {
   if (!isValid(url)) {
     let data = { errorMsg: 'Indirizzo URL invalido: ricompila il form' };
     return h.view('main-content', data, { layout: 'index' });
-
   }
 
   // Generate test SMTP service account from ethereal.email
@@ -53,7 +52,7 @@ module.exports = function (request, h) {
         referente: referente,
         url: url,
         amministrazione: amministrazione,
-        link: `http://localhost:3000/registered?token=${token}`
+        link: `http://localhost:3000/register-confirm?token=${token}`
       })
     };
 
@@ -74,9 +73,9 @@ module.exports = function (request, h) {
 
 /**
  * Controlla se l'URL e' ben formata e se il nome del dominio rientra in una whitelist di url.
- * 
+ *
  * @param url rappresenta l'url da controllare
- * 
+ *
  * @return true se la URL e' ben formata e se ientra in una whitelist, false altrimenti
  */
 function isValid(url) {
@@ -86,10 +85,10 @@ function isValid(url) {
 }
 
 /**
-* Controlla se il dominio della URL passata rientra in una whitelist di domini 
-* 
+* Controlla se il dominio della URL passata rientra in una whitelist di domini
+*
 * @param url rappresenta l'url da controllare
-* 
+*
 * @return true se l'url e' presente nella whitelist, false altrimenti
 */
 function isInWhiteList(url) {
@@ -125,11 +124,11 @@ function isInWhiteList(url) {
 
     let baseUrl = '';
 
-    /*  Gli url nella forma  https://try.gogs.io/<username>/<projectname>,  
-                            https://try.gitea.io/<username>/<projectname>, 
+    /*  Gli url nella forma  https://try.gogs.io/<username>/<projectname>,
+                            https://try.gitea.io/<username>/<projectname>,
                             https://secure.phabricator.com/project/view/395/
-        devono diventare    https://gogs.io/<username>/<projectname>,  
-                            https://gitea.io/<username>/<projectname>, 
+        devono diventare    https://gogs.io/<username>/<projectname>,
+                            https://gitea.io/<username>/<projectname>,
                             https://phabricator.com/project/view/<numProject>/
     */
     if (['gitea', 'gogs', 'phabricator'].includes(domain)) {
