+++ src/tpl/register-confirm.html
@@ -1,0 +1,29 @@
+<main id="content">
+    <div class="d-flex align-items-center justify-content-center main-content mt-5">
+        <div class="container">
+            <div class="row">
+                <div class="my-5 col-md-8">
+
+                    <h3>Confermare registrazione dell'URL nel database?</h3>
+
+                    <br/>
+                    Nome referente: {{referente}};
+                    <br/>
+                    Amministrazione richiedente: {{amministrazione}};
+                    <br/>
+                    Codice ipa: {{ipa}};
+                    <br/>
+                    Indirizzo PEC: {{pec}};
+                    <br/>
+                    URL di code hosting: {{url}};
+                    <br/>
+                    <br/>
+
+                    <a href="/registered?nomeReferente={{referente}}&ipa={{ipa}}&url={{url}}&pec={{pec}}&amministrazione={{amministrazione}}" class="btn btn-primary">Registra</a>
+                    <a href="/" class="btn btn-default">Annulla</a>
+
+                </div>
+            </div>
+        </div>
+    </div>
+</main>
